Added a way to track how many bytes are pending. This can then be compared against bytes copied in the manager.
It's a bit sticky because it relies on the lazy loading of file information that, for move operations, won't exist when we're looking to remove those pending bytes.

There's also the matter of special treatment. Though it makes sense that IOOperation should have a FileSize property - all files have sizes, regardless of what you're doing with them - it is still a bit strange that it's never used except for CopyOperations.

We also have a strange split in the tracking responsibility: OperationWorker holds it's own queue, so I figured it should hold the pending bytecount. But we're throwing away that information on garbage collection from dequeued items, and it's OperationManger's responsiblity to track things AFTER OperationWorker has done them (via the callback).

I need to consider that responsibility split some more.